Grant nginx Required Permissions
Using the latest Bookworm Raspberry Pi Lite (x64) OS and the Mainsail manual install guide will generate a 403 Forbidden error when attempting to access the printer.

This has been a missing step for several years, but I finally got around to submitting a pull request. There may be a better approach to grant nginx the file permissions it needs. But I'm not sure what that would be.